// apps/stats-service/src/enum/stats.param.names.enum.ts
enum StatsParamNameEnum {
  LEAD_LEADS = 'Leads',
  LEAD_CFTD = 'Cftd',
  LEAD_FTD = 'Ftd',
  LEAD_RETENTION = 'Retention',
  //
  TRANSFER_DEPOSIT = 'Deposit',
  TRANSFER_DEPOSIT_APPROVED = 'ApprovedDeposit',
  TRANSFER_DEPOSIT_REJECTED = 'RejectedDeposit',
  //
  TRANSFER_CREDIT = 'Credit',
  TRANSFER_CREDIT_APPROVED = 'ApprovedCredit',
  TRANSFER_CREDIT_REJECTED = 'RejectedCredit',
  //
  TRANSFER_CHARGEBACK = 'Chargeback',
  TRANSFER_CHARGEBACK_APPROVED = 'ApprovedChargeback',
  TRANSFER_CHARGEBACK_REJECTED = 'RejectedChargeback',
  //
  TRANSFER_WITHDRAWAL = 'Withdrawal',
  TRANSFER_WITHDRAWAL_APPROVED = 'ApprovedWithdrawal',
  TRANSFER_WITHDRAWAL_REJECTED = 'RejectedWithdrawal',
}

export default StatsParamNameEnum;